+++ src/php/Console/Infrastructure/Command/BuildCommands.php
@@ -7,14 +7,15 @@ namespace Phel\Console\Infrastructure\Command;
 use Phel\Build\Infrastructure\Command\BuildCommand;
 use Phel\Build\Infrastructure\Command\CacheClearCommand;
 use Phel\Console\Domain\ConsoleCommandProviderInterface;
+use Symfony\Component\Console\Command\LazyCommand;
 
 final class BuildCommands implements ConsoleCommandProviderInterface
 {
-    public function commands(): array
+    public function lazyCommands(): array
     {
         return [
-            new BuildCommand(),
-            new CacheClearCommand(),
+            new LazyCommand('build', ['b'], 'Build the current project', false, static fn(): BuildCommand => new BuildCommand()),
+            new LazyCommand('cache:clear', [], 'Clear the temp and cache directories', false, static fn(): CacheClearCommand => new CacheClearCommand()),
         ];
     }
 }