+++ src/php/Console/Infrastructure/Command/InteropCommands.php
@@ -6,13 +6,14 @@ namespace Phel\Console\Infrastructure\Command;
 
 use Phel\Console\Domain\ConsoleCommandProviderInterface;
 use Phel\Interop\Infrastructure\Command\ExportCommand;
+use Symfony\Component\Console\Command\LazyCommand;
 
 final class InteropCommands implements ConsoleCommandProviderInterface
 {
-    public function commands(): array
+    public function lazyCommands(): array
     {
         return [
-            new ExportCommand(),
+            new LazyCommand('export', [], 'Export all definitions with the meta data `{:export true}` as PHP classes', false, static fn(): ExportCommand => new ExportCommand()),
         ];
     }
 }