+++ src/php/Console/Infrastructure/Command/WatchCommands.php
@@ -6,13 +6,14 @@ namespace Phel\Console\Infrastructure\Command;
 
 use Phel\Console\Domain\ConsoleCommandProviderInterface;
 use Phel\Watch\Infrastructure\Command\WatchCommand;
+use Symfony\Component\Console\Command\LazyCommand;
 
 final class WatchCommands implements ConsoleCommandProviderInterface
 {
-    public function commands(): array
+    public function lazyCommands(): array
     {
         return [
-            new WatchCommand(),
+            new LazyCommand('watch', [], 'Watch Phel files and reload namespaces on change.', false, static fn(): WatchCommand => new WatchCommand()),
         ];
     }
 }