+++ src/php/Console/Infrastructure/ConsoleBootstrap.php
@@ -10,7 +10,6 @@ use Override;
 use Phel\Console\Application\WarnDeprecationsFlag;
 use Phel\Console\ConsoleFactory;
 use Symfony\Component\Console\Application;
-use Symfony\Component\Console\Command\Command;
 use Symfony\Component\Console\Input\ArgvInput;
 use Symfony\Component\Console\Input\InputInterface;
 use Symfony\Component\Console\Output\OutputInterface;
@@ -58,21 +57,6 @@ final class ConsoleBootstrap extends Application
         $this->getFactory()->getFilesystemFacade()->clearAll();
 
         exit($exitCode);
-    }
-
-    /**
-     * @return array<string,Command>
-     */
-    #[Override]
-    protected function getDefaultCommands(): array
-    {
-        $commands = parent::getDefaultCommands();
-
-        foreach ($this->getFactory()->getConsoleCommands() as $command) {
-            $commands[$command->getName()] = $command;
-        }
-
-        return $commands;
     }
 
     /**